Move profile type
